+++ vector/src/main/java/im/vector/app/features/home/room/detail/timeline/tools/EventRenderingTools.kt
@@ -26,12 +26,12 @@ import im.vector.app.core.utils.EvenBetterLinkMovementMethod
 import im.vector.app.core.utils.isValidUrl
 import im.vector.app.features.home.room.detail.timeline.TimelineEventController
 import im.vector.app.features.html.PillImageSpan
-import org.matrix.android.sdk.api.permalinks.MatrixLinkify
-import org.matrix.android.sdk.api.permalinks.MatrixPermalinkSpan
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
+import org.matrix.android.sdk.api.session.permalinks.MatrixLinkify
+import org.matrix.android.sdk.api.session.permalinks.MatrixPermalinkSpan
 
 fun CharSequence.findPillsAndProcess(scope: CoroutineScope, processBlock: (PillImageSpan) -> Unit) {
     scope.launch(Dispatchers.Main) {
